fix: refactor joi
