+++ src/validators/loginValidationSchema.mjs
@@ -1,0 +1,14 @@
+import Joi from "joi";
+
+const loginSchema = Joi.object({
+    email: Joi.string().email().required().messages({
+        "string.email": "A valid email is required.",
+        "any.required": "Email is required.",
+    }),
+    password: Joi.string().min(6).required().messages({
+        "string.min": "Password must be at least 6 characters long.",
+        "any.required": "Password is required.",
+    }),
+});
+
+export default loginSchema;
